refactor: createElement로 재사용성 높임
